requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.2.0
certifi==2024.7.4
charset-normalizer==3.3.2
comm==0.2.2
debugpy==1.8.2
decorator==5.1.1
exceptiongroup==1.2.1
executing==2.0.1
filelock==3.15.4
frozenlist==1.4.1
fsspec==2024.6.1
idna==3.7
ipykernel==6.29.5
ipython==8.26.0
jedi==0.19.1
jinja2==3.1.4
joblib==1.4.2
jupyter-client==8.6.2
jupyter-core==5.7.2
markupsafe==2.1.5
matplotlib-inline==0.1.7
mpmath==1.3.0
multidict==6.0.5
nest-asyncio==1.6.0
networkx==3.3
numpy==2.0.0
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.5.82
nvidia-nvtx-cu12==12.1.105
packaging==24.1
parso==0.8.4
pexpect==4.9.0
platformdirs==4.2.2
prompt-toolkit==3.0.47
psutil==6.0.0
ptyprocess==0.7.0
pure-eval==0.2.2
pyg-lib==0.4.0+pt23cu121
pygments==2.18.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pyzmq==26.0.3
requests==2.32.3
ruff==0.5.1
scikit-learn==1.5.1
scipy==1.14.0
six==1.16.0
stack-data==0.6.3
sympy==1.12.1
threadpoolctl==3.5.0
torch==2.3.1
torch-cluster==1.6.3+pt23cu121
torch-geometric==2.4.0
torch-scatter==2.1.2+pt23cu121
torch-sparse==0.6.18+pt23cu121
torch-spline-conv==1.2.2+pt23cu121
tornado==6.4.1
tqdm==4.66.4
traitlets==5.14.3
triton==2.3.1
typing-extensions==4.12.2
urllib3==2.2.2
wcwidth==0.2.13
yarl==1.9.4